+++ tests/performance/snr/mod.rs
@@ -1,6 +1,8 @@
 // Copyright (c) 2026 G & R Associates LLC
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+pub mod cofdm_fer;
+pub mod cofdm_stream_fer;
 pub mod dvb_t;
 pub mod ft8;
 pub mod ldpc_decode_rule;